Move async_rw_mutex to execution module
