+++ libs/pika/synchronization/tests/performance/CMakeLists.txt
@@ -1,22 +1,5 @@
-# Copyright (c) 2019 Hartmut Kaiser
+# Copyright (c) 2019 The STE||AR-Group
 #
 # SPDX-License-Identifier: BSL-1.0
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-set(benchmarks async_rw_mutex_scheduling)
-
-foreach(benchmark ${benchmarks})
-  set(sources ${benchmark}.cpp)
-
-  source_group("Source Files" FILES ${sources})
-
-  pika_add_executable(
-    ${benchmark}_test INTERNAL_FLAGS
-    SOURCES ${sources}
-    EXCLUDE_FROM_ALL ${${benchmark}_FLAGS}
-    FOLDER "Benchmarks/Modules/Synchronization"
-  )
-
-  pika_add_performance_test("modules.synchronization" ${benchmark} ${${benchmark}_PARAMETERS})
-endforeach()
