+++ libs/pika/synchronization/tests/unit/CMakeLists.txt
@@ -5,8 +5,6 @@
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 set(tests
-    async_rw_mutex
-    async_rw_mutex_yielding
     barrier
     binary_semaphore
     condition_variable
@@ -19,8 +17,6 @@ set(tests
     stop_token_cb2
 )
 
-set(async_rw_mutex_PARAMETERS THREADS 4)
-set(async_rw_mutex_yielding_PARAMETERS THREADS 4)
 set(barrier_cpp20_PARAMETERS THREADS 4)
 set(binary_semaphore_cpp20_PARAMETERS THREADS 4)
